Add ViewPager

+++ app/src/main/java/com/hamzacanbaz/weatherapp/data/model/weatherForecast/Coord.kt
@@ -1,6 +1,6 @@
 package com.hamzacanbaz.weatherapp.data.model.weatherForecast
 
 data class Coord(
-    val lat: Int,
+    val lat: Double,
     val lon: Double
 )
